requirements.txt
Flask==0.10.1
Flask-Assets==0.11
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-WTF==0.12
Jinja2==2.8
Mako==1.0.3
Markdown==2.6.5
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.11.2
bcrypt==2.0.0
blinker==1.4
cffi==1.3.1
flask-mongoengine==0.7.4
gunicorn==19.4.1
itsdangerous==0.24
misaka==2.0.0
mongoengine==0.10.5
passlib==1.6.5
py==1.4.30
pycparser==2.14
pymongo==3.2
pytest==2.8.5
python-editor==0.4
raven==5.8.1
requests==2.8.1
six==1.10.0
webassets==0.11.1
coverage==4.0.3
beautifulsoup4==4.4.1
mock==1.3.0
kombu==3.0.32
amqp==1.4.8
boto3==1.2.3
moto==0.4.19
pyyaml==3.11